moved logs summary into utils
