+++ backend/src/jobs/startSummaryScheduler.ts
@@ -1,14 +1,15 @@
 import cron from 'node-cron';
-import { updateLogSummary } from '../scripts/updateLogsSummary';
+import { updateLogSummary } from '../utils/updateLogsSummary.util';
+import logger from '../config/logger';
 
-console.log('[Cron] Starting log summary scheduler...');
+logger.debug('Starting log summary scheduler...');
 // Runs every 15 minutes (e.g. 00, 15, 30, 45 of each hour)
 cron.schedule('*/15 * * * *', async () => {
-  console.log('[Cron] Running log summary update every 15 mins...');
+  logger.debug('Running log summary update every 15 mins...');
   try {
     await updateLogSummary();
-    console.log('[Cron] Summary update completed');
+    logger.info('Summary update completed');
   } catch (err) {
-    console.error('[Cron] Error running summary update:', err);
+    logger.error('Error running summary update:', err);
   }
 });
